refactor(daily-fritz): split api.ts into clientStorage / apiErrors / api (R3)
`dailyFritz/api.ts` (891 LOC) carried three concerns: browser-storage cleanup,
the error type/classification/copy surface, and the actual network calls. Only
the last is "the Daily Fritz API".

- `dailyFritz/clientStorage.ts` — `DAILY_FRITZ_TODAY_CACHE_PREFIX` + the three
  `clear*` helpers.
- `dailyFritz/apiErrors.ts` — `resolveDailyFritzApiError`, the recoverable-code
  set + `isRecoverableDailyFritzAuthorityCode`, `DailyFritzAuthorityRecoveryError`,
  `throwDailyFritzAuthorityError`, `DailyFritzEndOfRunError`,
  `formatDailyFritzNextHandUserMessage`, `DailyFritzNextHandHttpError`,
  `isRetryableDailyFritzNextHandError`.
- `api.ts` keeps the request plumbing, response types, and the ~11 calls;
  imports what it needs from the two new modules.

`api.ts` re-exports the moved public symbols so all existing `dailyFritz/api`
consumers keep working unchanged — `useDailyFritzInit` and
`modules/daily/dailyFritzContracts` repointed at the split modules directly as
the example. Pure mechanical move, zero behaviour change.

REFACTOR_OPPORTUNITIES.md R3. `api.ts` drops 891→756 lines (still over the
`max-lines` budget — the API layer itself is large — so the warning count is
unchanged at 51, no new flags).

Local bar green: `tsc -b`, `lint` (51/51, 0 errors), `lint:hooks`,
`check:architecture` 20/20, `check:deps`, client `test:all` 218/1632, server
vitest all shards.
